requirements_v2.txt
google-generativeai
numpy
# For system agent code analysis
ast
time